fix(usage): restore predictions toggle to usage sidebar only
